+++ src/web.c
@@ -518,7 +518,7 @@ WebSetCommand(int ac, char *av[], void *arg)
   WebUser		u;
   int			port;
 
-  switch ((int) arg) {
+  switch ((intptr_t)arg) {
 
     case SET_OPEN:
       WebOpen(w);
